requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
playwright==1.42.0
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
pyyaml==6.0
pdfplumber==0.7.7
matplotlib==3.8.1
pytest==7.4.0
anyio==4.1.0
